Update prepared statements capability
The word 'client-side prepared statements' is a feature of PDO which will emulate prepared statements when the specific driver does not support. Because MySQL had supported prepared statements since version 4.1, PDO does not have to emulate them at all. So it's better not to tell reader about this, it can cause confusion (I saw a post in the fb group which a user thought mysqli does not support prepared statements).